+++ docs/app/components/PageHeader.vue
@@ -1,4 +1,7 @@
 <script lang="ts">
+/**
+ * @todo move to src
+ */
 export interface PageHeaderProps {
   title?: string
   description?: string
@@ -14,12 +17,7 @@ export interface PageHeaderSlots {
 </script>
 
 <script setup lang="ts">
-/**
- * @todo move to src
- */
-const props = withDefaults(defineProps<PageHeaderProps>(), {
-  title: 'horizontal' as const
-})
+const props = defineProps<PageHeaderProps>()
 const slots = defineSlots<PageHeaderSlots>()
 </script>
 
@@ -29,8 +27,10 @@ const slots = defineSlots<PageHeaderSlots>()
       <div class="flex flex-col items-start justify-between gap-[6px]">
         <div class="w-full flex flex-row items-center justify-between gap-[4px]">
           <div class="flex-1 max-w-[167px] sm:max-w-3/4">
-            <ProseH1 class="mb-0 truncate text-(--b24ui-typography-label-color) leading-(--ui-font-line-height-xs) font-(--ui-font-weight-light)">
-              {{ props.title }}
+            <ProseH1 v-if="props.title || !!slots.title" class="mb-0 truncate text-(--b24ui-typography-label-color) leading-(--ui-font-line-height-xs) font-(--ui-font-weight-light)">
+              <slot name="title">
+                {{ props.title }}
+              </slot>
             </ProseH1>
           </div>
           <div v-if="slots['head-links']" class="flex-1 flex flex-wrap flex-row items-center justify-end gap-[12px]">